+++ grpc_host/lib/src/configuration/ssl_settings.dart
@@ -10,12 +10,22 @@ class SslSettings {
   });
 
   factory SslSettings.fromYaml(YamlMap yaml) {
-    final certificatePath = yaml['certificatePath'];
-    final privateKeyPath = yaml['privateKeyPath'];
+    final certificatePath = yaml['certificatePath'] ?? '';
+    final privateKeyPath = yaml['privateKeyPath'] ?? '';
     final ret = SslSettings(
       certificatePath: certificatePath,
       privateKeyPath: privateKeyPath,
     );
     return ret;
+  }
+
+  SslSettings copyWith({
+    String? certificatePath,
+    String? privateKeyPath,
+  }) {
+    return SslSettings(
+      certificatePath: certificatePath ?? this.certificatePath,
+      privateKeyPath: privateKeyPath ?? this.privateKeyPath,
+    );
   }
 }