+++ grpc_host/CHANGELOG.md
@@ -1,7 +1,22 @@
+## 3.0.4
+
+- Support for empty SslSettings
+
+## 3.0.3
+
+- Added copyWith to SslSettings
+
+## 3.0.2
+
+- Added copyWith to HostSettings
+
+## 3.0.1
+
+- Deprecating smtpSettings
+
 ## 3.0.0
 
 - Split host settings from customizable AppSettings
-
 
 ## 2.0.0